auth: password validators
This is sort of a backport of Django 1.9's password validators.

The AUTH_PASSWORD_VALIDATORS allow is to enforce some minimum password
strength for setting user passwords. This allows us to be good citizens
and require stronger passwords on behalf of our users.

Right now, we're only enabling the MinimumLengthValidator, but we can
relatively figure out some slightly more complex algorithms or port some
of the other Django validator backends if we want to be fancier.